requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
python-dotenv==1.0.1
Flask-Mail==0.10.0